Add files via upload
youtube-dl is a command-line program to download videos from YouTube.com (https://github.com/rg3/youtube-dl/blob/master/README.md#readme). This is a script that downloads all of the latest video uploads on youtube channels that teach programing. The script will create a new folder for each channel and populate it with yesterday's uploads. You can specify the playlists/channels that you want to download from, the date (default is yesterday), and the maximum number of items in the playlist that you would like to iterate through (if you only care about videos downloaded yesterday then 10 should be enough for most channels). I find this very helpful for keeping me up to date on new programming videos.